requirements.txt
streamlit
PyPDF2
python-docx
nltk
scikit-learn
pandas
plotly
